requirements.txt
beautifulsoup4==4.11.1
fake_useragent==0.1.11
lxml==4.6.3
mysql_connector_repackaged==0.3.1
numpy==1.21.2
pandas==1.3.2
requests==2.26.0
streamlit==1.11.0
streamlit_on_Hover_tabs==1.0.1
streamlit_option_menu==0.3.2
urllib3==1.26.6
